Create ReferenceValidator interface and classes to implement it for AssayRef, StudyRef and SampleRef.
