+++ src/test/java/uk/ac/ebi/subs/validator/core/validators/SampleRefValidatorTest.java
@@ -16,9 +16,9 @@ import java.util.UUID;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class SampleRelationshipValidatorTest {
+public class SampleRefValidatorTest {
 
-    private ReferenceValidator validator;
+    private SampleRefValidator sampleRefValidator;
     private SampleRepository sampleRepository;
 
     private SampleRelationship relationship;
@@ -26,9 +26,9 @@ public class SampleRelationshipValidatorTest {
 
     @Before
     public void setUp() {
-        validator = new ReferenceValidator();
+        sampleRefValidator = new SampleRefValidator();
         sampleRepository = mock(SampleRepository.class);
-        validator.sampleRepository = sampleRepository;
+        sampleRefValidator.sampleRepository = sampleRepository;
 
         when(sampleRepository.findByAccession("SAMEA100001")).thenReturn(null);
         when(sampleRepository.findByAccession("SAMEA123456")).thenReturn(new Sample());
@@ -39,7 +39,7 @@ public class SampleRelationshipValidatorTest {
 
     @Test
     public void referenceNotFoundTest() {
-        validator.validateSampleRelationships(Arrays.asList(relationship), singleValidationResult);
+        sampleRefValidator.validateSampleRelationships(Arrays.asList(relationship), singleValidationResult);
         System.out.println(singleValidationResult.getMessage());
         Assert.assertEquals(ValidationStatus.Error, singleValidationResult.getValidationStatus());
     }
@@ -47,7 +47,7 @@ public class SampleRelationshipValidatorTest {
     @Test
     public void referenceFoundTest() {
         relationship.setAccession("SAMEA123456");
-        validator.validateSampleRelationships(Arrays.asList(relationship), singleValidationResult);
+        sampleRefValidator.validateSampleRelationships(Arrays.asList(relationship), singleValidationResult);
         System.out.println(singleValidationResult.getMessage());
 
         Assert.assertEquals(ValidationStatus.Pass, singleValidationResult.getValidationStatus());
